feat: styled pagefind cusotm ui
