+++ src/constants/transition.ts
@@ -1,0 +1,3 @@
+export const TITLE_TRANSITION_NAME = "title-transition";
+export const SITE_HEADER_TRANSITION_NAME = "site-header-transition";
+export const SEARCH_ICON_TRANSITION_NAME = "search-icon-transition";
