+++ src/features/_components/DarkModeButton.astro
@@ -7,26 +7,28 @@ const { class: className } = Astro.props;
 ---
 
 <button
-  id="theme-toggle"
-  class:list={["cursor-pointer", className]}
-  aria-label="Toggle theme"
+	id="theme-toggle"
+	class:list={["cursor-pointer", className]}
+	aria-label="Toggle theme"
 >
-  <slot />
+	<slot />
 </button>
 
 <script>
-  const handleToggleClick = () => {
-    const element = document.documentElement;
-    element.classList.toggle("dark");
+	const handleToggleClick = () => {
+		const element = document.documentElement;
+		element.classList.toggle("dark");
 
-    const isDark = element.classList.contains("dark");
-    localStorage.setItem("theme", isDark ? "dark" : "light");
-  }
+		const isDark = element.classList.contains("dark");
+		localStorage.setItem("theme", isDark ? "dark" : "light");
+	};
 
-  const initThemeToggleButton = () => {
-    document.getElementById("theme-toggle")?.addEventListener("click", handleToggleClick);
-  }
+	const initThemeToggleButton = () => {
+		document
+			.getElementById("theme-toggle")
+			?.addEventListener("click", handleToggleClick);
+	};
 
-  initThemeToggleButton();
-  document.addEventListener('astro:after-swap', initThemeToggleButton);
+	initThemeToggleButton();
+	document.addEventListener("astro:after-swap", initThemeToggleButton);
 </script>
