+++ src/features/_components/GoogleAnalytics.astro
@@ -9,6 +9,6 @@ const script = `window.dataLayer=window.dataLayer||[];function gtag(){dataLayer.
 ---
 
 <script
-  type="text/partytown"
-  src={`https://www.googletagmanager.com/gtag/js?id=${id}`}></script>
+	type="text/partytown"
+	src={`https://www.googletagmanager.com/gtag/js?id=${id}`}></script>
 <script type="text/partytown" set:html={script} />
