+++ src/features/posts/_components/BottomCard.astro
@@ -4,37 +4,36 @@ import { Icon } from "astro-icon/components";
 import Card from "./Card.astro";
 ---
 
-<Card
-  class="flex-col justify-between gap-4"
->
-  <div class="flex flex-col gap-3">
-    <div class="flex items-start gap-3">
-      <a href="https://ojii3.dev"
-        target="_blank"
-        rel="noopener noreferrer"
-        class="cursor-pointer duration-100 hover:rotate-30"
-        aria-label="ホームページへ"
-      >
-        <Image
-          src="/favicon-32x32.png" 
-          alt="icon"
-          class="w-8 h-8 object-cover rounded-full border-1 border-gray-2"
-          width={32}
-          height={32}
-        />
-      </a>
-      <a
-        href="https://github.com/OJII3/blog.ojii3.dev"
-        target="_blank"
-        rel="noopener noreferrer"
-        class="cursor-pointer duration-100 hover:rotate-30"
-        aria-label="GitHubリポジトリへ"
-      >
-        <Icon name="tabler:brand-github-filled" class="h-8 w-8" />
-      </a>
-    </div>
-  </div>
-  <p class="font-hachimaru text-xs leading-relaxed">
-    May you, the beauty of this world, always shine...
-  </p>
+<Card class="flex-col justify-between gap-4">
+	<div class="flex flex-col gap-3">
+		<div class="flex items-start gap-3">
+			<a
+				href="https://ojii3.dev"
+				target="_blank"
+				rel="noopener noreferrer"
+				class="cursor-pointer rounded-full duration-200 hover:rotate-30"
+				aria-label="ホームページへ"
+			>
+				<Image
+					src="/favicon-32x32.png"
+					alt="icon"
+					class="border-gray-2 h-8 w-8 rounded-full border-1 object-cover"
+					width={32}
+					height={32}
+				/>
+			</a>
+			<a
+				href="https://github.com/OJII3/blog.ojii3.dev"
+				target="_blank"
+				rel="noopener noreferrer"
+				class="cursor-pointer rounded-full duration-200 hover:rotate-30"
+				aria-label="GitHubリポジトリへ"
+			>
+				<Icon name="tabler:brand-github-filled" class="h-8 w-8" />
+			</a>
+		</div>
+	</div>
+	<p class="font-hachimaru text-xs leading-relaxed">
+		May you, the beauty of this world, always shine...
+	</p>
 </Card>
