+++ src/features/posts/_components/Card.astro
@@ -6,6 +6,11 @@ export interface Props {
 const { class: className } = Astro.props;
 ---
 
-<section class:list={["2x:w-[200px] 2x:h-[200px] flex h-[calc(50dvw-32px)] max-w-[200px] min-w-full  overflow-hidden rounded-2xl p-4 duration-200", className]}>
-  <slot />
+<section
+	class:list={[
+		"2x:w-[200px] 2x:h-[200px] 2x:max-w-[200px] flex h-[calc(50dvw-32px)] max-w-full min-w-full overflow-hidden rounded-2xl p-4 duration-200",
+		className,
+	]}
+>
+	<slot />
 </section>
