+++ src/features/posts/_components/PostPrevNext.astro
@@ -14,22 +14,18 @@ const { originalPost, targetPost, isPrev, isNext } = Astro.props;
 ---
 
 <a
-  href={`/${targetPost.id}`}
-  class:list={[
-    "flex w-full items-center gap-2 rounded-2xl border-2 p-4 duration-200",
-    BORDER_COLORS[originalPost.data.color],
-  ]}
-  data-astro-reload
+	href={`/${targetPost.id}`}
+	class:list={[
+		"flex w-full items-center gap-2 rounded-2xl border-2 p-4 duration-200",
+		BORDER_COLORS[originalPost.data.color],
+	]}
+	data-astro-reload
 >
-  {isPrev ? (
-    <Icon name="tabler:arrow-left" />
-  ) : null}
+	{isPrev ? <Icon name="tabler:arrow-left" /> : null}
 
-  <div class="flex-1">
-    <div class="text-sm">{targetPost.data.title}</div>
-  </div>
+	<div class="flex-1">
+		<div class="text-sm">{targetPost.data.title}</div>
+	</div>
 
-  {isNext ? (
-    <Icon name="tabler:arrow-right" />
-  ) : null}
+	{isNext ? <Icon name="tabler:arrow-right" /> : null}
 </a>
