+++ src/features/search/_components/SearchResult.astro
@@ -1,7 +1,9 @@
 ---
+
 ---
+
 <div
-  id="results"
-  class="w-full flex flex-col gap-4"
+	id="results"
+	class="3x:grid-cols-3 3x:w-3x 4x:grid-cols-4 4x:w-4x 5x:grid-cols-5 5x:w-5x 2x:w-2x 2x:columns-1x mx-auto grid max-w-[1096px] grid-cols-2 items-center justify-items-center gap-4 overflow-hidden"
 >
 </div>
