+++ src/pages/admin/index.astro
@@ -1,4 +1,5 @@
 ---
+
 ---
 
 This is the admin page.
