+++ src/pages/search.astro
@@ -1,57 +1,104 @@
 ---
+import { Icon } from "astro-icon/components";
+import { SITE_HEADER_TRANSITION_NAME } from "../constants";
 import GlobalLayout from "../features/_layouts/GlobalLayout.astro";
 import SearchForm from "../features/search/_components/SearchForm.astro";
 import SearchResult from "../features/search/_components/SearchResult.astro";
 ---
 
 <GlobalLayout title="Search">
-  <main class="max-w-[80ch] mx-auto p-4">
-    <div class="p-4">
-      <h1
-        transition:name=""
-        class="font-hachimaru 2x:text-xl flex text-lg font-bold"
-      >
-        晴れときどき崩壊ブログ
-      </h1>
-    </div>
-    <SearchForm />
-    <SearchResult />
-  </main>
-</GlobalLayout>
+	<main class="mx-auto max-w-[1096px] p-4">
+		<div class="flex flex-col gap-8">
+			<div
+				class="flex items-center gap-4 p-4"
+				transition:name={SITE_HEADER_TRANSITION_NAME}
+			>
+				<a
+					href="/"
+					class="group"
+					aria-label="Go back to home"
+					data-astro-history="auto"
+				>
+					<span
+						class="border-gray-5 flex h-12 w-12 items-center justify-center rounded-full border-2 p-2 duration-100"
+					>
+						<Icon name="tabler:arrow-left" class="h-6 w-6" />
+					</span>
+				</a>
+				<SearchForm />
+			</div>
 
-<script is:inline>
-const initSearch = () => {
-  document.querySelector("#search")?.addEventListener("input", async (e) => {
-    // Pagefindのスクリプトを最初の検索時にのみ読み込み
-    if (e.target.dataset.loaded !== "true") {
-      e.target.dataset.loaded = "true";
-      // Pagefindのスクリプトを読み込む
-      window.pagefind = await import("/pagefind/pagefind.js");
-    }
+			<SearchResult />
+		</div>
+	</main>
 
-    // 入力された値でインデックスを検索
-    const search = await window.pagefind.search(e.target.value);
+	<script is:inline>
+		const borderColors = [
+			"border-rose-500",
+			"border-orange-500",
+			"border-yellow-500",
+			"border-lime-500",
+			"border-emerald-500",
+			"border-indigo-500",
+			"border-purple-500",
+		];
+		const getIndexByDate = (dateString) => {
+			const date = new Date(dateString);
+			return date.getTime() % borderColors.length;
+		};
 
-    // 以前の検索結果をクリア
-    document.querySelector("#results").innerHTML = "";
+		const initSearch = () => {
+			const searchInput = document.querySelector("#search");
+			const resultsContainer = document.querySelector("#results");
 
-    // 新しい検索結果を追加
-    for (const result of search.results) {
-      const data = await result.data();
-      document.querySelector("#results").innerHTML += `
+			searchInput.addEventListener("input", async (e) => {
+				// Pagefindのスクリプトを最初の検索時にのみ読み込み
+				if (e.target.dataset.loaded !== "true") {
+					e.target.dataset.loaded = "true";
+					window.pagefind = await import("/pagefind/pagefind.js");
+				}
+
+				const search = await window.pagefind.search(e.target.value);
+
+				// 以前の検索結果をクリア
+				resultsContainer.innerHTML = "";
+
+				// 新しい検索結果を追加
+				for (const result of search.results) {
+					const data = await result.data();
+					const borderColor = borderColors[getIndexByDate(data.meta.date)];
+					document.querySelector("#results").innerHTML += `
 <a
-  class="relative flex flex-col"
+  class="2x:w-[200px] 2x:h-[200px] flex flex-col h-[calc(50dvw-32px)] max-w-full 2x:max-w-[200px] min-w-full overflow-hidden rounded-2xl p-4 gap-2 border-2 ${borderColor}"
   href="${data.url}"
 >
-  <h3 class="-mb-3 z-1">
-    <span class="bg-gray-0 font-bold px-2 ml-2">${data.meta.title}</span>
-  </h3>
-  <p class="border-2 rounded-2xl p-4">${data.excerpt}</p>
+  <h3 class="text-gray-5 text-xs">${data.meta.title} </h3>
+  <p class="overflow-hidden">
+    <span class="overflow-ellipsis text-sm font-bold">${data.excerpt}</span>
+  </p>
 </a>`;
-    }
-  });
-};
+				}
+			});
 
-document.addEventListener("astro:after-swap", initSearch);
-initSearch();
-</script>
+			// 確定時に入力値をクエリパラメータに反映
+			searchInput.closest("form")?.addEventListener("submit", (e) => {
+				e.preventDefault();
+				const query = searchInput.value;
+				const url = new URL(window.location);
+				url.searchParams.set("search", query);
+				window.history.replaceState(null, "", url);
+			});
+
+			// URLのクエリパラメータからデフォルト検索語を設定
+			const urlParams = new URLSearchParams(window.location.search);
+			const query = urlParams.get("search");
+			if (query && searchInput.value !== query) {
+				searchInput.value = query;
+				searchInput.dispatchEvent(new Event("input"));
+			}
+		};
+
+		document.addEventListener("astro:after-swap", initSearch);
+		initSearch();
+	</script>
+</GlobalLayout>
